riscv: Add regalloc_push/pop/prepare_move utilities
This commit adds `regalloc_push`, `regalloc_pop` and `regalloc_prepare_move` to
the `Assembler`.

`regalloc_push` pushes a value (either in a register or in the `JITFrame`
(`r.jfp`)) to the native stack (`r.sp`). `regalloc_pop` pops a value from the
stack.  These two functions are used by `remap_frame_layout_mixed` when it
wants to spill a value to break circular dependency.

`regalloc_prepare_move` loads a value from the stack slot of `JITFrame` to the
register when both `src` and `dst` are stack slots.  This is needed for the
RISC-V backend because RISC-V ISA doesn't support memory-to-memory move.