riscv: Emit large int or float in constant pool
This commit changes how we load large integer immediates
(`imm < -SINT32_MIN or imm > SINT32_MAX`) or float immediates.

For large integer immediates, instead of building a full 64-bit integer
with a sequence of `lui + addiw + slli + addi` (up to 8 instructions),
we simply emit `auipc + ld` when the immediate can't be represented as a
32-bit signed integer.  For 32-bit signed integer, we continue using
`lui + addiw`.

For float immediates, instead of saving the immediate value in
`MachineDataBlockWrapper`, loading address with `load_int_imm`, and
loading the final float value with a `load_float`, we emit the float
immediate in the constant pool and emit a pc-relative `load_float`.
This saves a word for address and two instructions to load address.
This also avoids pointer chasing.