riscv: Support large frame slot offsets
This commit adds support for large frame slot offsets.  Before this
commit, the code generator assumes the frame slot offsets is always
addressable with signed 12-bit immediate.  But, it turns out 12-bit is
not enough.  `lib_python_tests.py` `test_tarfile` triggers the assertion
in the code generator.

This commit fixes the problem by emitting `load_imm + ADD` when the
immediate is too large for `ADDI`.

Test: pypy2.7 testrunner/lib_python_tests.py -v -k test_tarfile