extra_tests: Fix test_sqlite3 test_connection_del
This commit fixes the failing `test_connection_del` test case in
`test_sqlite3.py`.

The test case rely on `gc.collect()` to invoke the `Connection.__del__`
method.  But the `con` list in the `open_many` function is still
reachable after the first `open_many(False)` invocation (Note: it seems
that the exception holds a backtrace, the backtrace holds the frame
object, the frame object keeps the `con` list the `Connection` objects
reachable.)

This commit fixes the problem by wrapping the first invocation in a
function so that the propagating exception object can be freed before we
call `gc.collect()`.